Changed buttons to menu view, minor improvements

+++ WindowsFormsApp1/RequestObject.cs
@@ -5,11 +5,16 @@ namespace ReqRaportsApp
     public class request
     {
         public string clientId { get; set; }
-        public int requestId { get; set; }
+        public long requestId { get; set; }
         public string name { get; set; }
         public int quantity { get; set; }
         public double price { get; set; }
     }
+
+    //public class RequestList
+    //{
+    //    public List<request> RequestsList = new List<request>();
+    //}
 
     public class WholeReq
     {
